Refactoring the U-Net 3D model so that it will be easier to create a integration test for the application.
