+++ applications/segmentation/unet3d/unet3d_network_architectures.py
@@ -1,11 +1,6 @@
-import argparse
-
 import lbann
 import lbann.models
-import lbann.contrib.args
-import lbann.contrib.launcher
 import lbann.modules as lm
-from lbann.core.util import get_parallel_strategy_args
 
 
 class UNet3D(lm.Module):
@@ -192,161 +187,3 @@ class Deconvolution3dModule(lm.ConvolutionModule):
 
     def __init__(self, *args, **kwargs):
         super().__init__(3, transpose=True, *args, **kwargs)
-
-
-def create_unet3d_data_reader(train_dir, test_dir):
-    readers = []
-    for role, shuffle, role_dir in [
-            ("train", True, train_dir),
-            ("test", False, test_dir)]:
-        if role_dir is None:
-            continue
-
-        readers.append(lbann.reader_pb2.Reader(
-            name="hdf5",
-            role=role,
-            shuffle=shuffle,
-            data_file_pattern="{}/*.hdf5".format(role_dir),
-            validation_percent=0,
-            percent_of_data_to_use=1.0,
-            scaling_factor_int16=1.0,
-            hdf5_key_data="volume",
-            hdf5_key_labels="segmentation",
-            hdf5_hyperslab_labels=True,
-            disable_labels=False,
-            disable_responses=True,
-        ))
-
-    return lbann.reader_pb2.DataReader(reader=readers)
-
-
-def create_unet3d_optimizer(learn_rate):
-    # TODO: This is a temporal optimizer copied from CosomoFlow.
-    adam = lbann.Adam(
-        learn_rate=learn_rate,
-        beta1=0.9,
-        beta2=0.999,
-        eps=1e-8)
-    return adam
-
-
-if __name__ == '__main__':
-    desc = ('Construct and run the 3D U-Net on a 3D segmentation dataset.'
-            'Running the experiment is only supported on LC systems.')
-    parser = argparse.ArgumentParser(description=desc)
-    lbann.contrib.args.add_scheduler_arguments(parser)
-
-    # General arguments
-    parser.add_argument(
-        '--job-name', action='store', default='lbann_unet3d', type=str,
-        help='scheduler job name (default: lbann_unet3d)')
-    parser.add_argument(
-        '--mini-batch-size', action='store', default=1, type=int,
-        help='mini-batch size (default: 1)', metavar='NUM')
-    parser.add_argument(
-        '--num-epochs', action='store', default=5, type=int,
-        help='number of epochs (default: 100)', metavar='NUM')
-
-    # Model specific arguments
-    parser.add_argument(
-        '--learning-rate', action='store', default=0.001, type=float,
-        help='the initial learning rate (default: 0.001)')
-    parser.add_argument(
-        '--partition-level', action='store', default=4, type=int,
-        help='the spatial partition level (default: 4)')
-    parser.add_argument(
-        '--depth-groups', action='store', type=int, default=4,
-        help='the number of processes for the depth dimension (default: 4)')
-    default_lc_dataset = '/p/vast1/lbann/datasets/LiTS/hdf5_dim128_float'
-    default_train_dir = '{}/train'.format(default_lc_dataset)
-    default_test_dir = '{}/test'.format(default_lc_dataset)
-    parser.add_argument(
-        '--train-dir', action='store', type=str, default=default_train_dir,
-        help='the directory of the training dataset (default: \'{}\')'
-        .format(default_train_dir))
-    parser.add_argument(
-        '--test-dir', action='store', type=str, default=default_test_dir,
-        help='the directory of the test dataset (default: \'{}\')'
-        .format(default_test_dir))
-
-    parser.add_argument(
-        '--dynamically-reclaim-error-signals', action='store_true',
-        help='Allow LBANN to reclaim error signals buffers (default: False)')
-
-    parser.add_argument(
-        '--batch-job', action='store_true',
-        help='Run as a batch job (default: false)')
-
-    lbann.contrib.args.add_optimizer_arguments(
-        parser,
-        default_optimizer="adam",
-        default_learning_rate=0.001,
-    )
-    args = parser.parse_args()
-
-    parallel_strategy = get_parallel_strategy_args(
-        sample_groups=args.mini_batch_size,
-        depth_groups=args.depth_groups)
-
-    # Construct layer graph
-    volume = lbann.Input(data_field='samples')
-    segmentation = lbann.Input(data_field='label_reconstruction')
-    output = UNet3D()(volume)
-    ce = lbann.CrossEntropy(
-        [output, segmentation],
-        use_labels=True)
-    obj = lbann.ObjectiveFunction([ce])
-    layers = list(lbann.traverse_layer_graph([volume, segmentation]))
-    for l in layers:
-        l.parallel_strategy = parallel_strategy
-
-    # Setup model
-    metrics = [lbann.Metric(ce, name='CE', unit='')]
-    callbacks = [
-        lbann.CallbackPrint(),
-        lbann.CallbackTimer(),
-        lbann.CallbackGPUMemoryUsage(),
-        lbann.CallbackProfiler(skip_init=True),
-    ]
-    # # TODO: Use polynomial learning rate decay (https://github.com/LLNL/lbann/issues/1581)
-    # callbacks.append(
-    #     lbann.CallbackPolyLearningRate(
-    #         power=1.0,
-    #         num_epochs=100,
-    #         end_lr=1e-5))
-    model = lbann.Model(
-        epochs=args.num_epochs,
-        layers=layers,
-        objective_function=obj,
-        callbacks=callbacks,
-    )
-
-    # Setup optimizer
-    optimizer = lbann.contrib.args.create_optimizer(args)
-
-    # Setup data reader
-    data_reader = create_unet3d_data_reader(
-        train_dir=args.train_dir,
-        test_dir=args.test_dir)
-
-    # Setup trainer
-    trainer = lbann.Trainer(mini_batch_size=args.mini_batch_size)
-
-    # Runtime parameters/arguments
-    environment = lbann.contrib.args.get_distconv_environment(
-        num_io_partitions=args.depth_groups)
-    if args.dynamically_reclaim_error_signals:
-        environment['LBANN_KEEP_ERROR_SIGNALS'] = 0
-    else:
-        environment['LBANN_KEEP_ERROR_SIGNALS'] = 1
-    lbann_args = ['--use_data_store']
-
-    # Run experiment
-    kwargs = lbann.contrib.args.get_scheduler_kwargs(args)
-    lbann.contrib.launcher.run(
-        trainer, model, data_reader, optimizer,
-        job_name=args.job_name,
-        environment=environment,
-        lbann_args=lbann_args,
-        batch_job=args.batch_job,
-        **kwargs)
