+++ python/lbann/contrib/args.py
@@ -77,9 +77,11 @@ def get_distconv_environment(parallel_io=False, num_io_partitions=1):
 
     return {
         'DISTCONV_WS_CAPACITY_FACTOR': 0.8,
-        'DISTCONV_OVERLAP_HALO_EXCHANGE': 1,
-        'LBANN_DISTCONV_HALO_EXCHANGE': 'HYBRID',
-        'LBANN_DISTCONV_TENSOR_SHUFFLER': 'HYBRID',
+#        'DISTCONV_OVERLAP_HALO_EXCHANGE': 1,
+        'LBANN_DISTCONV_HALO_EXCHANGE': 'AL',
+        'LBANN_DISTCONV_TENSOR_SHUFFLER': 'AL',
+#        'LBANN_DISTCONV_HALO_EXCHANGE': 'HYBRID',
+#        'LBANN_DISTCONV_TENSOR_SHUFFLER': 'HYBRID',
         'LBANN_DISTCONV_CONVOLUTION_FWD_ALGORITHM': 'AUTOTUNE',
         'LBANN_DISTCONV_CONVOLUTION_BWD_DATA_ALGORITHM': 'AUTOTUNE',
         'LBANN_DISTCONV_CONVOLUTION_BWD_FILTER_ALGORITHM': 'AUTOTUNE',
